feat(ui): complete mobile UI redesign and QR scanner fixes

+++ td-backend/src/main/java/com/lichiapps/techdepot/controllers/RefTipoContenedorController.java
@@ -14,5 +14,5 @@ public class RefTipoContenedorController {
     
     @GetMapping public List<RefTipoContenedor> getAll() { return repository.findAll(); }
     @PostMapping public RefTipoContenedor save(@RequestBody RefTipoContenedor entity) { return repository.save(entity); }
-    @DeleteMapping("/{id}") public void delete(@PathVariable Integer id) { repository.deleteById(id); }
+    @DeleteMapping("/{id}") public void delete(@PathVariable Long id) { repository.deleteById(id); }
 }
